// client/model/panosApplicationFirewallRuleDto.ts
/**
 * Draas
 * Draas OAS Documentation
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PanosApplication } from './panosApplication';


export interface PanosApplicationFirewallRuleDto { 
    firewallRuleId: string;
    panosApplicationsToAdd: Array<PanosApplication>;
    panosApplicationsToRemove: Array<PanosApplication>;
}